Renamed and refactored tests
